pyproject.toml
[tool.poetry]
name = "kxm-helper"
version = "0.1.0"
description = ""
authors = ["ivanm <ivanmuranov595@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.6.0"
aiogram-dialog = "^2.1.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
openai = "^1.30.3"
redis = "^5.0.4"
sulguk = "^0.8.0"
fluent-runtime = "^0.4.0"
markdown = "^3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
